完成了察看据点信息页厂商类型据点关联游戏的页面。 退出登录时应为服务器返回非 500 错误。

+++ app/root-controller.js
@@ -17,10 +17,12 @@
                 }).then(function (response) {
                     union.$localStorage.user = response.data;
                     _czc.push(["_setCustomVar", "登录用户", response.data.IdCode + "-" + response.data.UserName, 1]);
-                }, function () {
-                    $http.delete(apiEndpoint + "login/current");
-                    delete union.$localStorage.login;
-                    notification.error("登录失效，请重新登录。");
+                }, function (response) {
+                    if(response.status !== 500){
+                        $http.delete(apiEndpoint + "login/current");
+                        delete union.$localStorage.login;
+                        notification.error("登录失效，请重新登录。");
+                    }
                 });
             }
